[microphone_loopback] add E70/AK4954, E54/WM8904 and E54/WM8904+FreeRTOS variations MH3-9343 MH3-2248 MH3-9163

+++ apps/microphone_loopback/firmware/src/config/sam_e70_xult_wm8904_ssc_wqvga/gfx/libaria/src/libaria_widget_bar_graph_skin_classic.c
@@ -120,7 +120,7 @@ static void drawTickLabelWithValue(laBarGraphWidget* graph, GFX_Point tickPoint,
     //Protect from overflow
     if (value < MAX_TICK_LABEL_VALUE)
     {
-        sprintf(strbuff, "%ld", value);
+        sprintf(strbuff, "%d", (int) value);
     }
     else
     {
